new beta version 0.8.3
this version contains the result of the refactoring of the plugin "com.lorands.hunspell4eclipse.hunspelljna" :
- Uses the bundles dependances between the hunspell plugins (no import packages)
- Reuses the classes available in the project "ikea"
- Includes the JNA lib 3.2.7  and the native libs in special folders (validated on Mac Os X1.6.7 64bits, Windows XP SP3 32bits and Ubuntu 10.04 LTS 64bits)
- Corrects the method "libName()" in the class "Hunspell"

Includes a logging tool which can be controlled with a system property ( -Dhunspell.log.on ) . 
When this tool is explicitly activated (debug mode), it writes the lines in the console, in the other case it stores the lines in the eclipse logging tool.
Each line contains 6 parts: hh:mm:ss.sss, level, thread name,instance-id,method,infos

Uses a "java.text.BreakIterator" to split the text of the regions in the method "checkInner()".  Done to manage the words which include one or more digit.

Adds a "locale" property in the class "Dictionary" to be able to instanciate the BreakIterator with the locale of the current dictionary

Corrects of the  exclusion rule "WWMixedCaseIgnored" in the method "checkRules()" "AbstractHunSpellEngine" 

Corrects the extensions declared in the plugin "com.lorands.hunspell4eclipse.javaSpell" .
Eclipse showed error dialog boxes when it tried to instanciate the unexisting class associated to the extension point "javaCompletionProposalComputer"